http, pcspeaker driver
Former-commit-id: a3ad38695e8c1e1040a6a3f4b8470217e8a98489
Former-commit-id: cd3809edcae7b26e7b3d846dab17f1f63761f30f

+++ src/net/torvald/terrarum/virtualcomputer/terminal/Terminal.kt
@@ -61,7 +61,7 @@ interface Terminal : Teletype {
      * @param duration: milliseconds
      * @param freg: Frequency (float)
      */
-    fun emitTone(duration: Int, freq: Float)
+    fun emitTone(duration: Int, freq: Double)
 
     override fun bell(pattern: String)
     /** Requires keyPressed() event to be processed.
